#90837 Move Session object to Service
Functionally it is more a Service than an Object, also restore Address
